+++ src/ashley/machina_extensions/forum_permission/handler.py
@@ -8,6 +8,9 @@ from django.db import models
 from machina.apps.forum_permission.handler import (
     PermissionHandler as BasePermissionHandler,
 )
+from machina.core.db.models import get_model
+
+Forum = get_model("forum", "Forum")
 
 
 class PermissionHandler(BasePermissionHandler):
@@ -23,11 +26,11 @@ class PermissionHandler(BasePermissionHandler):
         self.current_lti_context_id: Optional[int] = None
 
     def can_archive_forum(self, forum, user):
-        """ Given a forum, checks whether the user can archive it. """
+        """Given a forum, checks whether the user can archive it."""
         return self._perform_basic_permission_check(forum, user, "can_archive_forum")
 
     def can_rename_forum(self, forum, user):
-        """ Given a forum, checks whether the user can rename it. """
+        """Given a forum, checks whether the user can rename it."""
         return self._perform_basic_permission_check(forum, user, "can_rename_forum")
 
     def can_manage_moderator(self, forum, user):
@@ -76,3 +79,16 @@ class PermissionHandler(BasePermissionHandler):
                 )
             )
         return readable_forums
+
+    # pylint:disable = W0201
+    def _get_all_forums(self):
+        """Return all forums accessible for the LTIContext of the user."""
+        if not hasattr(self, "_all_forums"):
+            if self.current_lti_context_id:
+                self._all_forums = list(
+                    Forum.objects.filter(lti_contexts__id=self.current_lti_context_id)
+                )
+            else:
+                super()._get_all_forums()
+
+        return self._all_forums