+++ CHANGELOG.md
@@ -10,13 +10,14 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- track forum views, topic and post updates and creations with xAPI events
 - allow to search users by any of its meaningful fields in Django admin
 - allow to archive a forum with the new `can_archive_forum` permission
 - track topic views with XAPI events
 - automatically assign a public_username to instructors and administrators
-when none is defined in the LTI authentication
-- allow  administrators and instructors to move topics to forums in the same 
-LTIContext
+  when none is defined in the LTI authentication
+- allow administrators and instructors to move topics to forums in the same
+  LTIContext
 
 ## [1.0.0] - 2021-08-16